Update changes from review
