+++ kingfisher_scrapy/spiders/pakistan_ppra_releases.py
@@ -1,9 +1,12 @@
+import json
+
 import scrapy
 
 from kingfisher_scrapy.base_spider import SimpleSpider
+from kingfisher_scrapy.util import handle_http_error, components
 
 
-class PackistanPPRAReleases(SimpleSpider):
+class PakistanPPRAReleases(SimpleSpider):
     """
     Domain
       Pakistan Public Procurement Regulatory Authority (PPRA)
@@ -14,4 +17,15 @@ class PackistanPPRAReleases(SimpleSpider):
     data_type = 'release_package'
 
     def start_requests(self):
-        yield scrapy.Request('https://www.ppra.org.pk/api/index.php/api/release', meta={'file_name': 'releases.json'})
+        yield scrapy.Request(
+            'https://www.ppra.org.pk/api/index.php/api/records',
+            meta={'file_name': 'list.html'},
+            callback=self.parse_list
+        )
+
+    @handle_http_error
+    def parse_list(self, response):
+        # remove the last item in the list to fix the str JSON format
+        urls = json.loads(response.xpath('//body//text()').getall()[6].replace(",\r\n\r\nhttps://www.ppra.org.pk", ""))
+        for url in urls:
+            yield self.build_request(url, formatter=components(-2))
